+++ packages/pdf-viewer/src/PDFViewer.tsx
@@ -32,7 +32,7 @@ function mergeArrays<T, U>(arr1: T[][] = [], arr2: U[][] = []): (T | U)[][] {
   return result;
 }
 
-const defaultLoadingTips = <div className="pdf-viewer-default-empty-tips">正在打开文件...</div>;
+const defaultLoadingTips = <div className="pdf-viewer-default-loading-tips">正在打开文件...</div>;
 
 const defaultEmptyTips = <div className="pdf-viewer-default-empty-tips">请打开一个 PDF 文件</div>;
 
@@ -65,6 +65,9 @@ export interface PDFViewerProps extends Omit<React.HTMLAttributes<HTMLDivElement
   /** 空文件提示 */
   emptyTips?: React.ReactElement;
 
+  /** 自定义加载过程提示 */
+  loadingTips?: React.ReactElement;
+
   /** 标注内容变化事件 */
   onMarkChange?: PDFPainterPluginProps['onMarkChange'];
 
@@ -93,6 +96,7 @@ const PDFViewer = React.forwardRef<PDFViewerHandle, PDFViewerProps>((props, pRef
     onPageScroll,
     children,
     emptyTips = defaultEmptyTips,
+    loadingTips = defaultLoadingTips,
     title: _title,
     hideToolbar,
     defaultTitle,
@@ -265,12 +269,13 @@ const PDFViewer = React.forwardRef<PDFViewerHandle, PDFViewerProps>((props, pRef
     setPages([]);
   });
   const load = useMemoizedFn<PDFViewerHandle['load']>(async (file) => {
-    let pdfContent = file;
-    if (pdfContent instanceof File) {
-      pdfContent = await pdfContent.arrayBuffer();
-    }
     if (pdfJs) {
       setLoading(true);
+      let pdfContent = file;
+      if (pdfContent instanceof File) {
+        pdfContent = await pdfContent.arrayBuffer();
+      }
+
       try {
         const pdfDoc = await pdfJs.getDocument(pdfContent).promise;
         if (pdfDoc) {
@@ -593,34 +598,37 @@ const PDFViewer = React.forwardRef<PDFViewerHandle, PDFViewerProps>((props, pRef
               setToolbarRightDom(dom);
             }}
           />
-          <div
-            ref={pageContainerRef}
-            className={styles.pages}
-            onScroll={onPageScroll}
-            style={{
-              // @ts-expect-error
-              '--scale-factor': scale,
-              '--pdf-viewer-page-scale': scale,
-            }}
-          >
-            {viewports.length === 0 && (loading ? defaultLoadingTips : emptyTips)}
-            {viewports.map((viewport, pageIndex) => {
-              const shouldRender = pageIndex >= renderRange[0] && pageIndex <= renderRange[1];
-              const width = `calc(var(--scale-factor) * ${Math.floor(viewport.width)}px)`;
-              const height = `calc(var(--scale-factor) * ${Math.floor(viewport.height)}px)`;
-              const gap = `calc(var(--scale-factor) * ${pageGap}px)`;
-              return (
-                <div key={pageIndex} className={styles.pageContainer} style={{ width, height, marginBottom: gap }}>
-                  {shouldRender && (
-                    <>
-                      <PDFPage className={styles.page} index={pageIndex} zoom={zoom} render={shouldRender} />
-                      <div ref={node => (pageCoverRefs[pageIndex] = node)} className={styles.pageCover} />
-                      <div className={styles.pageCover}>{renderPageCover(pageIndex, { viewport, zoom })}</div>
-                    </>
-                  )}
-                </div>
-              );
-            })}
+          <div className={styles.pagesOuter}>
+            <div
+              ref={pageContainerRef}
+              className={styles.pages}
+              onScroll={onPageScroll}
+              style={{
+                // @ts-expect-error
+                '--scale-factor': scale,
+                '--pdf-viewer-page-scale': scale,
+              }}
+            >
+              {viewports.length === 0 && !loading && emptyTips}
+              {viewports.map((viewport, pageIndex) => {
+                const shouldRender = pageIndex >= renderRange[0] && pageIndex <= renderRange[1];
+                const width = `calc(var(--scale-factor) * ${Math.floor(viewport.width)}px)`;
+                const height = `calc(var(--scale-factor) * ${Math.floor(viewport.height)}px)`;
+                const gap = `calc(var(--scale-factor) * ${pageGap}px)`;
+                return (
+                  <div key={pageIndex} className={styles.pageContainer} style={{ width, height, marginBottom: gap }}>
+                    {shouldRender && (
+                      <>
+                        <PDFPage className={styles.page} index={pageIndex} zoom={zoom} render={shouldRender} />
+                        <div ref={node => (pageCoverRefs[pageIndex] = node)} className={styles.pageCover} />
+                        <div className={styles.pageCover}>{renderPageCover(pageIndex, { viewport, zoom })}</div>
+                      </>
+                    )}
+                  </div>
+                );
+              })}
+            </div>
+            {loading && loadingTips}
 
             {/* 绘图的工具栏渲染 */}
             {!disabledPluginPainter && <PDFPainterPlugin ref={pdfPainterHandle} onMarkChange={onMarkChange} />}