+++ packages/pocket/src/img/Img.style.ts
@@ -1,0 +1,38 @@
+import jssAutoPrefix from '@orca-fe/jss-plugin-auto-prefix';
+import { createUseStyles } from '../utils/jss';
+
+const prefix = 'orca-img';
+
+export default createUseStyles(
+  {
+    root: {
+      position: 'relative',
+    },
+    loading: {
+      position: 'absolute',
+      top: '0',
+      left: '0',
+      width: '100%',
+      height: '100%',
+    },
+    '@global': {
+      '.orca-img-error-tip': {
+        position: 'absolute',
+        top: '0',
+        left: '0',
+        width: '100%',
+        height: '100%',
+        display: 'flex',
+        justifyContent: 'center',
+        alignItems: 'center',
+        backgroundColor: 'rgba(125, 125, 125, 0.1)',
+        textShadow: '0 0 1px #fff',
+        color: '#999999',
+      },
+    },
+  },
+  {
+    classNamePrefix: prefix,
+    plugins: [jssAutoPrefix({ prefix })],
+  },
+);