+++ packages/transformer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @orca-fe/painter
 
+## 0.0.10
+
+### Patch Changes
+
+- feat: transformer 完善文档 & 添加 onDataChange 导出参数
+
+## 0.0.9
+
+### Patch Changes
+
+- feat: PDFViewer 自定义 loading 效果
+
 ## 0.0.8
 
 ### Patch Changes